FINNNNNNNNNNNNNNN

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: mqueguin <mqueguin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/10 14:46:07 by mqueguin          #+#    #+#             */
-/*   Updated: 2021/07/06 13:32:12 by mqueguin         ###   ########.fr       */
+/*   Updated: 2021/08/06 16:33:27 by mqueguin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int		i;
 	int		negative;
